requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21
pandas>=1.3
Pillow>=9.0
pyyaml
tqdm
nltk
tensorboard
wandb
fastapi
uvicorn[standard]
python-multipart
pydantic
gradio
onnxruntime
pytest
